LPS-60613 move
